I don't even know what I'm doing anymore

+++ crates/defaults/src/resources/defaults/shaders/others/default_impls/general.func.glsl
@@ -1,4 +1,4 @@
 // General info
-uniform double _time;
-uniform double _delta;
+uniform float _time;
+uniform float _delta;
 uniform ivec2 _resolution;
